requirements.txt
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
arviz==0.11.2
asn1crypto==1.0.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cryptography==2.7
decorator==5.0.6
defusedxml==0.7.1
depinfo==1.7.0
diskcache==5.2.1
entrypoints==0.3
future==0.18.2
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
importlib-metadata==3.10.0
importlib-resources==5.1.2
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.6.0
jupyterlab==3.0.13
jupyterlab-launcher==0.13.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mpi4py==3.0.0
mpmath==1.2.1
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
optlang==1.5.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
pymc==2.3.8
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-libsbml==5.19.0
python-libsbml-experimental==5.18.1
pytz==2021.1
PyYAML==5.3
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
rfc3986==1.4.0
rich==6.2.0
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
sniffio==1.2.0
sphinx==4.1.2
swiglpk==5.0.3
sympy==1.7.1
terminado==0.9.4
testpath==0.4.4
tornado==6.1
tqdm==4.36.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
coverage==6.5.0
pytest-cov==4.0.0
